a few more JPAResource functions
