--- asm/JSystem/JParticle/JPAResource/__ct__11JPAResourceFv.s
@@ -1,29 +1,0 @@
-lbl_80274010:
-/* 80274010  38 00 00 00 */	li r0, 0
-/* 80274014  90 03 00 08 */	stw r0, 8(r3)
-/* 80274018  90 03 00 04 */	stw r0, 4(r3)
-/* 8027401C  90 03 00 00 */	stw r0, 0(r3)
-/* 80274020  90 03 00 18 */	stw r0, 0x18(r3)
-/* 80274024  90 03 00 14 */	stw r0, 0x14(r3)
-/* 80274028  90 03 00 10 */	stw r0, 0x10(r3)
-/* 8027402C  90 03 00 0C */	stw r0, 0xc(r3)
-/* 80274030  90 03 00 1C */	stw r0, 0x1c(r3)
-/* 80274034  90 03 00 20 */	stw r0, 0x20(r3)
-/* 80274038  90 03 00 24 */	stw r0, 0x24(r3)
-/* 8027403C  90 03 00 28 */	stw r0, 0x28(r3)
-/* 80274040  90 03 00 2C */	stw r0, 0x2c(r3)
-/* 80274044  90 03 00 30 */	stw r0, 0x30(r3)
-/* 80274048  90 03 00 34 */	stw r0, 0x34(r3)
-/* 8027404C  90 03 00 38 */	stw r0, 0x38(r3)
-/* 80274050  98 03 00 47 */	stb r0, 0x47(r3)
-/* 80274054  98 03 00 46 */	stb r0, 0x46(r3)
-/* 80274058  98 03 00 45 */	stb r0, 0x45(r3)
-/* 8027405C  98 03 00 44 */	stb r0, 0x44(r3)
-/* 80274060  98 03 00 43 */	stb r0, 0x43(r3)
-/* 80274064  98 03 00 42 */	stb r0, 0x42(r3)
-/* 80274068  98 03 00 41 */	stb r0, 0x41(r3)
-/* 8027406C  98 03 00 40 */	stb r0, 0x40(r3)
-/* 80274070  98 03 00 3F */	stb r0, 0x3f(r3)
-/* 80274074  98 03 00 3E */	stb r0, 0x3e(r3)
-/* 80274078  B0 03 00 3C */	sth r0, 0x3c(r3)
-/* 8027407C  4E 80 00 20 */	blr
